Feat/admin

+++ src/main/java/com/project/zipmin/repository/LikeRepository.java
@@ -12,6 +12,6 @@ public interface LikeRepository extends JpaRepository<Like, Integer> {
 	
     boolean existsByTablenameAndRecodenumAndUserId(String tablename, int recodenum, int userId);
 
-	long countByTablenameAndRecodenum(String tablename, int recodenum);
+	int countByTablenameAndRecodenum(String tablename, int recodenum);
 	
 }
